requirements.txt
django>=3.2,<4.0
gunicorn
requests
django-dotenv
psycopg2-binary
django-storages
boto3
djangorestframework
django-cleanup
pillow